Expose EventEmitter for Events. Expose retry opts.
Exposing an RxJS observable for the public API of the events prevents
the user from potentially wrapping the EventEmitter with their own
abstraction of choice. It also blocks access to the 'disconnect' and
'reconnect' events, which may be useful information for the user.

This patch exposes the Event Emitter of the ReconnectingWebSocket
directly. Closes #3.

This patch also allows full customization of the options passed to
`retry.operation()` by passing `retryOptions` into
`awry.Events.connect()` or if using ReconnectingWebSocket directly,
passing the same into the constructor. Closes #2.