Modify source to align with standard/prettier
Most of the changes are moving from single to double quotes. Method
chains were also affected. I'm not in love with the Prettier method
chains, but there's never been a defacto "best" way to do it so i'll
live with that. Lastly, the standard rule 'no-callback-literal' was
triggering on the verifyClient callback of the websocket server, so i
renamed that callback method to something that wouldn't trigger the
rule.