Use eslint-plugin-prettier
Previously we used `prettier --list-different` to output files that
weren't formatted with prettier. This required a hacky npm script to do
the check and print a reasonable error message to the user.

This patch switches from the npm script approach to integrating with
eslint using a plugin. This allows eslint to report any formatting
differences as a lint failure, which improves the reporting and
readability of the error output.