Convert to ecmascript modules; Add esdoc
This patch moves the library to using ecmascript modules supported by
the @std/esm library[1]. This change is motivated by the desire to
provide solid and thorough library API documentation in a browsable
format.

By converting the library to use ecmascript modules, it created
a happy path for using the esdoc[2] documentation generation library.
esdoc is by far the best tool for the job at the moment, providing the
best quality documentation "out of the box".

This patch is a major version bump, because it renames many properties
on objects that were previously technically part of the public API, but
not advertised as such. These properties include:

- "request" and "baseUrl" on all API classes (i.e.
client.applications.request, client.sounds.baseUrl).

- "ws", "wsOptions", "retryOptions", "url", "reconnect" properties on
the ReconnectingWebSocket class (`returned by awry.Events.connect()`)

Fixes #6.
[1]: https://github.com/standard-things/esm
[2]: https://github.com/esdoc/esdoc