Update to ws@5, remove now-unneeded event code
When ws@4 was released, we had to put some code in the close() method of
the ReconnectingWebSocket to prevent error events from bubbling up
during the process of shutting down a websocket connection. In ws@5,
this is no longer necessary, as these events no longer occur.

This is a breaking change, see
https://github.com/websockets/ws/releases/tag/5.0.0 for the breaking
changes in the underlying `ws` library.

Closes #63.
Closes #64.