Stop using mjs file extension
As of esm@3.0.57, .mjs files are no longer allowed to be "side-loaded"
using a plain require call. As a result, mocha was no longer able to
load the tests.

This patch renames all the .mjs files back to .js and regenerates the
documentation to match the new filenames. Tooling config has also been
updated.

"use strict" was also removed from the tests as it is unneeded in an
esmodule.

Closes #79.